Add custom group intercept to reset indentation
This allows me to reset the `console.group()` indentation should a method fail
and not cleanup its grouping -- helps with back-to-back calls of `queryServer`
in a development environment (test suite).  It is being used exclusively here
and not in the production codebase because it would be more than irresponsible
to override global functions in a library.

This commit note is way too long -- I'm just trying to justify this hackery.